chore: re-organize types
