+++ lib/src/types/types.dart
@@ -17,6 +17,7 @@ export 'images.dart';
 export 'lat_lng.dart';
 export 'lat_lng_bounds.dart';
 export 'markers.dart';
+export 'navigation.dart';
 export 'navigation_destinations.dart';
 export 'navigation_initialization_params.dart';
 export 'navigation_view_types.dart';
